--- packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/features/ExposureOffsetValue.java
@@ -1,17 +1,0 @@
-package io.flutter.plugins.camera.features;
-
-/**
- * This represents the exposure offset value. It holds the minimum and
- * maximum values, as well as the current setting value.
- */
-public class ExposureOffsetValue {
-    final public double min;
-    final public double max;
-    final public double value;
-
-    public ExposureOffsetValue(double min, double max, double value) {
-        this.min = min;
-        this.max = max;
-        this.value = value;
-    }
-}
