+++ packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/features/Point.java
@@ -1,14 +1,12 @@
 package io.flutter.plugins.camera.features;
 
-/**
- * Represents a point on an x/y axis.
- */
+/** Represents a point on an x/y axis. */
 public class Point {
-    final public Double x;
-    final public Double y;
+  public final Double x;
+  public final Double y;
 
-    public Point(Double x, Double y) {
-        this.x = x;
-        this.y = y;
-    }
+  public Point(Double x, Double y) {
+    this.x = x;
+    this.y = y;
+  }
 }
